Cleanup the code that does filament changes so it's not crazy.   Also make it better.
Changes a few things here.  The filament change now won't nag you with
any extra buttons or required clicks.  You start the change, and when
you put the new filament in you hit 1 button on the LCD and it just does
it.  When it's done it beeps so you know it's done, but you don't have
to touch the mahcine or anything at all.  When it's done it's done and
ready to go.

The print head also sits high up enough and in the middle of the build
plate to make grabbing the purge really easy.